backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2026.1.4
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.4.4
colorama==0.4.6
distro==1.9.0
daphne==4.1.2
dj-database-url==2.1.0
Django==6.0.1
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.4.0
django-ratelimit==4.1.0
django-redis==5.4.0
redis==5.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
groq==0.15.0
google-generativeai>=0.5.0
google-genai>=1.0.0
google-auth==2.27.0
openai==2.15.0
pillow==12.1.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.1
sqlparse==0.5.5
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yt-dlp==2025.12.8

# Production WSGI server
gunicorn==23.0.0

# Error tracking
sentry-sdk==2.22.0

# Static files serving
whitenoise==6.8.2

# Cloud storage for media files (Railway has ephemeral filesystem)
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# Admin theme
django-jazzmin==3.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Security
django-ratelimit==4.1.0
googlesearch-python
duckduckgo-search>=7.0.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
markdown==3.7
sendgrid==6.11.0
feedparser==6.0.11
beautifulsoup4==4.14.3

# Vector Database & Semantic Search
# Using FAISS instead of ChromaDB (no C++ compiler needed)
langchain==0.3.15
langchain-google-genai==2.0.8
langchain-community==0.3.15
faiss-cpu==1.9.0.post1

# ML Quality Scorer
scikit-learn>=1.5.0
joblib>=1.4.0
